+++ EssentialFeed/FeedAPI/RemoteFeedLoader.swift
@@ -8,16 +8,6 @@
 
 import Foundation
 
-public enum HTTPClientResult {
-    case success(Data, HTTPURLResponse)
-    case failure(Error)
-}
-
-public protocol HttpClient {
-    func get(from url:URL, completion: @escaping (HTTPClientResult) -> Void)
-}
-
-
 public final class RemoteFeedLoader {
     private let url: URL
     private let client:HttpClient
